# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Task Management API
  version: 0.0.1
  description: Backend API documentation
servers:
  - url: http://localhost:5050/api
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Attachment:
      type: object
      properties:
        id: { type: string, format: uuid }
        filename: { type: string }
        url:
          {
            type: string,
            description: 'Relative file path under UPLOAD_DIR',
          }
        mimeType:
          { type: string, description: 'e.g., image/png, application/pdf' }
        size: { type: integer, description: 'File size in bytes' }
        issueId: { type: string, format: uuid }
        uploadedBy: { type: string, format: uuid }
        createdAtUTC: { type: string, format: date-time }
        type: { type: string, enum: [IMAGE, DOCUMENT, VIDEO, OTHER] }
        hasInlinePreview: { type: boolean }
    ErrorResponse:
      type: object
      properties:
        error: { type: string }

    # ==========================================
    # INPUT MODELS
    # ==========================================
...
paths:
...
  /issues/{issueId}/attachments:
    get:
      summary: Get all attachments for an issue
      tags: [Attachments]
      parameters:
        - in: path
          name: issueId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: A list of attachments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Attachment' }
    post:
      summary: Upload one or more attachments to an issue
      tags: [Attachments]
      parameters:
        - in: path
          name: issueId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: One or more files (repeat "file" field)
      responses:
        '201':
          description: Attachment(s) uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Attachment' }

  # ==========================================
  # ATTACHMENTS
  # ==========================================
  /attachments/{attachmentId}:
    get:
      summary: Get attachment metadata
      tags: [Attachments]
      parameters:
        - in: path
          name: attachmentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Attachment metadata (including the URL to download/view it)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Attachment' }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete an attachment
      tags: [Attachments]
      parameters:
        - in: path
          name: attachmentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Attachment deleted successfully
...